Revert "chore!: components big rename (#50)"
This reverts commit 3ce265831717c89daa8204680f0ba45fd43c0404.
