+++ src/components/document/ArticleNode.vue
