+++ src/components/document/property-fields/PropertyAnyOf.spec.ts
