+++ src/components/document/property-fields/PropertyDescription.vue
