+++ src/components/document/property-fields/PropertyEnum.spec.ts
