+++ src/components/document/property-fields/PropertyEnum.vue
