+++ src/components/document/property-fields/PropertyExample.spec.ts
