+++ src/components/document/property-fields/PropertyExample.vue
