+++ src/components/document/property-fields/PropertyInfo.spec.ts
