+++ src/components/document/property-fields/PropertyInfo.vue
