+++ src/components/document/property-fields/PropertyPattern.spec.ts
