+++ src/components/document/property-fields/PropertyPattern.vue
