+++ src/components/document/property-fields/PropertyRange.spec.ts
