+++ src/components/document/property-fields/PropertyRange.vue
