+++ src/components/table-of-contents/DividerItem.vue
