+++ src/components/table-of-contents/ExternalLinkItem.vue
