+++ src/components/table-of-contents/GroupItem.vue
