+++ src/components/table-of-contents/NodeGroupItem.vue
