+++ src/components/table-of-contents/NodeItem.vue
