+++ src/components/table-of-contents/index.ts
